requirements.txt
# FastAPI 및 웹 프레임워크
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# 데이터베이스
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# 인증 및 보안
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# 설정 관리
pydantic-settings==2.0.3
python-dotenv==1.0.0

# HTTP 클라이언트
httpx==0.25.2
aiofiles==23.2.1

# 유틸리티
email-validator==2.1.0
python-dateutil==2.8.2

# 개발 도구
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
isort==5.12.0
flake8==6.1.0

# 로깅
structlog==23.2.0

# 캐싱 (선택사항)
redis==5.0.1

# AWS S3 (선택사항)
boto3==1.34.0
